ats-service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pdfminer.six
python-docx
scikit-learn
numpy
sentence-transformers
nltk

